Extract base styles
